requirements.txt
torch>=1.13.0
torchaudio>=0.13.0
numpy>=1.20.0
demucs>=4.0.0
pyyaml>=6.0
# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.1.0
unittest-mock>=1.3.0 